[Fix] 충돌예방
