+++ TUDY/Models/User.swift
@@ -61,6 +61,16 @@ struct User: Codable {
     }
 }
 
+extension Encodable {
+    /// Object to Dictionary
+    /// cf) Dictionary to Object: JSONDecoder().decode(Object.self, from: dictionary)
+    var asDictionary: [String: Any]? {
+        guard let object = try? JSONEncoder().encode(self),
+              let dictinoary = try? JSONSerialization.jsonObject(with: object, options: []) as? [String: Any] else { return nil }
+        return dictinoary
+    }
+}
+
 struct UserForRegister {
     static var shared: User = User()
     private init() {}
