--- TUDY/Util/Extension/Codable+Custom.swift
@@ -1,18 +1,0 @@
-//
-//  Codable+Custom.swift
-//  TUDY
-//
-//  Created by neuli on 2022/06/15.
-//
-
-import Foundation
-
-extension Encodable {
-    /// Object to Dictionary
-    /// cf) Dictionary to Object: JSONDecoder().decode(Object.self, from: dictionary)
-    var asDictionary: [String: Any]? {
-        guard let object = try? JSONEncoder().encode(self),
-              let dictinoary = try? JSONSerialization.jsonObject(with: object, options: []) as? [String: Any] else { return nil }
-        return dictinoary
-    }
-}
